Feature: Applications Settings
- Cleanup
- Updated documentation
- Reset configuration functionality
